% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{AnnHAge}
\alias{AnnHAge}
\title{Households distribution by age in Annecy urban area in 2011}
\format{A Spatial object including 19 polygons corresponding to each municipality of Grenoble Urban Area (Insee definition) and following data attributes:}
\source{
Insee: Resultats du recensement de la population 2011, \href{http://www.insee.fr/}{Insee}
}
\usage{
data(AnnHAge)
}
\description{
This data set gives the households distribution by age in
Annecy urban area in 2011, including the area vectorial map at
municipality level.
}
\details{
\itemize{
  \item code: Municipality code
  \item name: Municipality name
  \item young: 15-29 years old
  \item adult: 30-59 years old
  \item old: 60 years or more
}
}
\keyword{datasets}

