% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{GreHSize}
\alias{GreHSize}
\title{Households distribution by  size in Grenoble urban area in 2011}
\format{
A Spatial object including 52 polygons corresponding to each municipality of Grenoble Urban Area (Insee definition) and following data attributes:
}
\source{
Insee: Resultats du recensement de la population 2011, \href{https://www.insee.fr/fr/accueil}{Insee}
}
\usage{
data(GreHSize)
}
\description{
This data set gives the households distribution by size in 
Grenoble urban area in 2011, including the area vectorial map at
municipality level.
}
\details{
\itemize{
  \item code: Municipality code
  \item name: Municipality name
  \item small: (1-2 persons household)
  \item medium: (3-4 persons household)
  \item big: (more then 5 persons household)
}
}
\keyword{datasets}
