% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_model.R
\docType{data}
\name{nopd_oasis_model}
\alias{nopd_oasis_model}
\title{OASIS model without PD}
\format{
A \code{glm} object
}
\usage{
nopd_oasis_model
}
\description{
A \code{glm} object containing the OASIS model 
from the updated, corrected
data, fit without the PD data.
}
\keyword{datasets}
