% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_model.R
\docType{data}
\name{not2_oasis_model}
\alias{not2_oasis_model}
\title{OASIS model without T2}
\format{A \code{glm} object}
\usage{
not2_oasis_model
}
\description{
A \code{glm} object containing the OASIS model 
from the updated, corrected
data, fit without the T2 data.
}
\keyword{datasets}
