% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{segdata}
\alias{segdata}
\title{Theoretical two groups distribution on a 10x10 grid map.}
\format{
A 10x10 grid Spatial object and following data attributes:
}
\source{
Hong S-Y (2014): R package 'seg', \href{https://sites.google.com/site/hongseongyun/seg/}{Hong's page}
}
\usage{
data(segdata)
}
\description{
The theoretical examples ( Morrill 1991, Wong 1993) adapted 
from Hong and O'Sullivan (2015). The space is represented by 
a 10x10 checkboard, with different distributions of two social 
groups in the area.
}
\details{
\itemize{
   \item : spatial ID;
   \item : municipality name;
   \item : pattern A: minority distribution;
   \item : pattern A: majority distribution;
   \item : pattern B: minority distribution;
   \item : pattern B: majority distribution;
   \item : pattern C: minority distribution;
   \item : pattern C: majority distribution;
   \item : pattern D: minority distribution;
   \item : pattern D: majority distribution;
   \item : pattern E: minority distribution;
   \item : pattern E: majority distribution;
   \item : pattern F: minority distribution;
   \item : pattern F: majority distribution;
   \item : pattern G: minority distribution;
   \item : pattern G: majority distribution;
   \item : pattern H: minority distribution;
   \item : pattern H: majority distribution;
   \item : pattern I: minority distribution;
   \item : pattern I: majority distribution;
}
}
\keyword{datasets}
