% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegFunctions.R
\name{SP}
\alias{SP}
\title{A function to compute the spatial proximity index (SP)}
\usage{
SP(x, d = NULL, fdist = 'e', distin = 'm',  distout = 'm', diagval = '0', 
itype = 'multi', beta = 1, spatobj = NULL, folder = NULL, shape = NULL)
}
\arguments{
\item{x}{- an object of class matrix (or which can be coerced to that class), 
where each column represents the distribution of a group within 
spatial units. The number of columns should be greater than 1 (at least 2 
groups are required). You should not include a column with total 
population, because this will be interpreted as a group.}

\item{d}{- a matrix of the distances between spatial unit centroids}

\item{fdist}{- the method used for distance interaction matrix: 
e' for inverse exponential function (by default) and 'l' for linear.}

\item{distin}{- input metric conversion, based on  \pkg{bink} package and 
includes conversions from 'm', 'km', 'inch', 'ft', 'yd', 'mi', 'naut_mi', etc.}

\item{distout}{- output metric conversion, based on  \pkg{bink} package and 
includes conversions to 'm', 'km', 'inch', 'ft', 'yd', 'mi', 'naut_mi', etc.}

\item{diagval}{- when providing a spatial object or a shape file, 
the user has the choice of the spatial matrix diagonal definition: 
diagval = '0' (by default) for an null diagonal and diagval = 'a' 
to compute the diagonal as 0.6 * square root (spatial/organizational unitsarea) (White, 1983)}

\item{itype}{- a character string defining the index type:
itype = 'multi' (by default) for the multi-group index (White, 1986),
itype = 'between' for the between groups version (White, 1983), or
itype = 'one' for the one-group version (Apparicio et al, 2008)}

\item{beta}{- distance decay parameter}

\item{spatobj}{- a spatial object (SpatialPolygonsDataFrame) with 
geographic information}

\item{folder}{- a character vector with the folder (directory) 
name indicating where the shapefile is located on the drive}

\item{shape}{- a character vector with the name of the shapefile 
(without the .shp extension).}
}
\value{
spatial proximity index value(s)
}
\description{
The spatial proximity index, SP, compares the clustering 
level (mean proximity) of a group compared to another group. 
The function can be used in two ways: to provide a distance matrix 
or a external geographic information source (spatial object 
or shape file).
}
\examples{
x <- segdata@data[ ,1:2]
ar<-area(segdata)
dist <- distance(segdata)
diag(dist)<-sqrt(ar) * 0.6
foldername <- system.file('extdata', package = 'OasisR')
shapename <- 'segdata'

SP(x, spatobj = segdata)

SP(x, folder = foldername, shape = shapename, fdist = 'l', itype = 'between') 

SP(x, spatobj = segdata, diagval ='a', itype = 'one')

SP(x, d = dist, fdist = 'e')

}
\references{
White M. J. (1983) \emph{The Measurement of Spatial 
Segregation}. American Journal of Sociology, 88, p. 1008-1019.

White, M. J. (1986) \emph{Segregation and Diversity Measures 
in Population Distribution}E. Population Index 52(2): 198-221.

Apparicio, P., V. Petkevitch and M. Charron (2008): \emph{Segregation 
Analyzer: A C#.Net application for calculating residential segregation indices}, 
Cybergeo: European Journal of Geography, 414, 1-27.
}
\seealso{
Proximity measures: \code{\link{Pxx}}, 
\code{\link{Pxy}},  \code{\link{Poo}}

Clustering Indices: 
\code{\link{ACL}}, \code{\link{RCL}}
}
